fix parser for openbor

+++ src/openbor/data.cpp
@@ -675,7 +675,7 @@ std::string * combineItems(const Value & items){
 std::string * toString(const Value & input){
     std::ostringstream out;
     for (Value::iterator it = input.getValues().begin(); it != input.getValues().end(); it++){
-        out << (char) (long) (*it).getValue();
+        out << (char) (intptr_t) (*it).getValue();
     }
     std::string * object = makeString(out.str());
     return object;
